Move VR to valuerep
