+++ pydicom/tests/test_charset.py
@@ -4,12 +4,13 @@
 
 import pytest
 
-import pydicom.charset
 from pydicom import dcmread, config
+import pydicom.charset
+from pydicom.charset import DEFAULT_CHARSET, CUSTOMIZABLE_CHARSET
 from pydicom.data import get_charset_files, get_testdata_file
 from pydicom.dataelem import DataElement
 from pydicom.filebase import DicomBytesIO
-from pydicom.valuerep import PersonName
+from pydicom.valuerep import PersonName, STR_VR
 
 # The file names (without '.dcm' extension) of most of the character test
 # files, together with the respective decoded PatientName tag values.
@@ -498,3 +499,8 @@ class TestCharset:
                                 "s in encoded string"):
             encoded = pydicom.charset.encode_string('あaｱア', ['shift_jis'])
             assert b'?a??' == encoded
+
+
+def test_vr_configured():
+    """Test that all str-like VRs have their charsets configured"""
+    assert STR_VR == DEFAULT_CHARSET | CUSTOMIZABLE_CHARSET
