+++ pydicom/util/leanread.py
@@ -12,7 +12,7 @@ from pydicom.misc import size_in_bytes
 from pydicom.datadict import dictionary_VR
 from pydicom.tag import TupleTag, ItemTag
 from pydicom.uid import UID
-from pydicom.vr import EXPLICIT_VR_LENGTH_32
+from pydicom.valuerep import EXPLICIT_VR_LENGTH_32
 
 
 extra_length_VRs_b = tuple(vr.encode('ascii') for vr in EXPLICIT_VR_LENGTH_32)
